ext4: kunit tests for extent splitting and conversion
Add multiple KUnit tests to test various permutations of extent
splitting and conversion.

We test the following cases:

1. Split of unwritten extent into 2 parts and convert 1 part to written
2. Split of unwritten extent into 3 parts and convert 1 part to written
3. Split of unwritten extent into 2 unwritten extents
4. Split of unwritten extent into 3 unwritten extents
5. Split of written extent into 2 parts and convert 1 part to unwritten
6. Split of written extent into 3 parts and convert 1 part to unwritten
7. Zeroout fallback for all the above cases

The main function we test here is ext4_split_convert_extents().
Currently some of the tests are failing due to issues in implementation.
Some failures are actually rare but valid failures [1], whereas some are
mitigated at other layers in ext4 [2] but still point out the mismatch in
expectation of what the caller wants vs what the function does.

The aim is to eventually fix all the failures we see here.

[1] for example, zeroout fallback exposes stale data in written to
unwritten conversion.

[2] for example, EXT4_GET_BLOCKS_CONVERT doesn't
really convert the split extent to written, but rather the callers end up
doing the conversion.

Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>