ext4: Allow zeroout when doing written to unwritten split
Currently, when we are doing an extent split and convert operation
of written to unwritten extent (example, done by ZERO_RANGE), we
don't allow the zeroout fallback incase the extent tree manipulation
fails.

Now, that we have refactored zeroout to be efficient and resepect
max_zeroout limit, we should be able to zeroout relatively quickly.
Hence, alllow zeroout for this case as well, so we can leverage the
fallback.

Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>